+++ analysis/tests/test.sh
@@ -1,4 +1,4 @@
-for file in src/*.{res,resi}; do
+for file in document/src/*.{res,resi}; do
   output="$(dirname $file)/expected/$(basename $file).txt"
   ../rescript-editor-analysis.exe test $file &> $output
   # CI. We use LF, and the CI OCaml fork prints CRLF. Convert.
@@ -7,7 +7,7 @@ for file in src/*.{res,resi}; do
   fi
 done
 
-for file in not_compiled/*.res; do
+for file in document/not_compiled/*.res; do
   output="$(dirname $file)/expected/$(basename $file).txt"
   ../rescript-editor-analysis.exe test $file &> $output
   # CI. We use LF, and the CI OCaml fork prints CRLF. Convert.
@@ -15,6 +15,19 @@ for file in not_compiled/*.res; do
     perl -pi -e 's/\r\n/\n/g' -- $output
   fi
 done
+
+# Workspaces tests
+for dir in workspaces/*/; do
+  testname="$(basename $dir)"
+  output="$(dirname $dir)/$testname/expected/$testname.txt"
+  dirtest="$(dirname $dir)/$testname"
+  ../rescript-editor-analysis.exe test $dirtest &> $output
+  # CI. We use LF, and the CI OCaml fork prints CRLF. Convert.
+  if [ "$RUNNER_OS" == "Windows" ]; then
+    perl -pi -e 's/\r\n/\n/g' -- $output
+  fi
+done
+
 
 warningYellow='\033[0;33m'
 successGreen='\033[0;32m'